restructure scroll-sync and sticky header portal
