+++ src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/components/helpers/scroll_sync/index.tsx
