+++ src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/components/helpers/sticky_header_portal/index.tsx
